Gave DataUser its own file
- Pushed import of Pint down into Quantity use for better initial import
  time
