+++ PyOpenWorm/configure.py
@@ -31,8 +31,11 @@ class Configure(object):
     # conf: is a configure instance to base this one on
     # dependencies are required for this class to be initialized (TODO)
 
-    def __init__(self):
-        self._properties = {}
+    def __init__(self, **kwargs):
+        for x in kwargs:
+            if not isinstance(kwargs[x],ConfigValue):
+                kwargs[x] = _C(kwargs[x])
+        self._properties = kwargs
 
     def __setitem__(self, pname, value):
         if not isinstance(value, ConfigValue):
